pyproject.toml
[project]
name = "AetherGIS"
version = "2.0.0"
description = "AI-Based Temporal Enhancement & WebGIS Visualization System — Production Grade GeoAI Platform"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    # HTTP & WMS
    "httpx>=0.27.0",
    "owslib>=0.29.0",
    "aiofiles>=23.2.1",
    # Image processing
    "opencv-python>=4.9.0.80",
    "pillow>=10.3.0",
    "numpy>=1.26.4",
    "scikit-image>=0.23.2",
    "matplotlib>=3.9.0",          # MODULE 6/7 — heatmap rendering
    # AI / ML
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    # Video generation
    "ffmpeg-python>=0.2.0",
    # Job queue (MODULE 1)
    "celery[redis]>=5.4.0",
    "redis>=5.0.4",
    "kombu>=5.3.5",
    # Configuration & validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.1",
    # Logging
    "structlog>=24.1.0",
    # Storage
    "boto3>=1.34.0",
    "imageio-ffmpeg>=0.6.0",
    # Performance monitoring (MODULE 14)
    "psutil>=5.9.0",
    "pynvml>=11.5.0",             # GPU utilisation via NVML
    # SSE streaming (MODULE 11) — via fastapi built-in StreamingResponse
    # Rate limiting (MODULE 12) — pure Python / Redis
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "pytest-cov>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.hatch.build.targets.wheel]
packages = ["backend"]


